Add initial collection of custom functions
Add four custom (compared to Go's standard strings package [1])
functions over strings.

The first function is named `All` and tests whether or not every string
in a slice satisfies some condition. The behaviour is mostly intuitive.
The return value for an empty slice is inspired by a JavaScript method
for arrays called "every" [2], which returns true for empty arrays.

The second function is named `Any` and tests whether or not any string
in a slice satisfies some condition. The behaviour is mostly intuitive.
The return value for an empty slice is inspired by a JavaScript method
for arrays called "some" [3], which returns false for empty arrays.

The third function is named `isEmpty` and tests whether or not a string
is the empty string [4]. That is, whether or not it is equal to `""`.

The fourth and final function is named `MapAll` and maps a slice of
strings according to a mapping function. The name `MapAll` is necessary
since there is already a `Map` function in the standard strings package
[1].

--
1. https://golang.org/pkg/strings
2. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/every
3. https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some
4. https://en.wikipedia.org/wiki/Empty_string