Add isValidUTF8 function (#1)
* Add IsValidUTF8 function

Add a function called IsValidUTF8 that checks if the provided string is
a valid UTF-8 string. I.e., if the string contains only characters that
are in the UTF-8 character set.

* Test IsValidUTF8 function

Add tests for the (new) IsValidUTF8 function. One testable example and
one slightly more extensive unit test suite.

* Update README

Add the (new) IsValidUTF8 function to the list of extra functions in the
README.